Updated mod2mp3 to use openmpt123 and do a bit more error checking
Amazingly, neither xmp nor openmpt123 exit with an error code if they
can't recognize a module file, or even if given a file that doesn't
exist. But they will still happily output a bullshit .wav file with
nothing in it!

We now run an openmpt123 --probe on the file before attempting to
convert it. Which also doesn't exit with an error status, so we have to
scrape the output using grep. What a fucking crock.