Rework of nextrack and recode for proper gapless playback
Under the previous iteration, tracks that belong together could be
played as a group. A classic example (not uncommonly heard on classic
rock stations) is 'Another Brick in the Wall Part II' preceeded by Part
I and 'Happiest Years Of Their Lives'.

There were two issues occuring:

1) There was a short silent gap in between tracks, as the Madplay
   decoder isn't able to handle gapless playback.
2) Track volumes would jump in between tracks. This is because the audio
   (at implemented in the KT4L project) were being run through sox
   with a dynamnic range compressor filter. Since the tracks were being
   decoded and filtered individually, when a new track was played, a new
   sox instance was called, with no knowledge of the state of the
   previous process, and therefore resetting the volume

The solution was:

1) Use mpg123 instead of madplay, which can handle gapless playback
2) Pass the tracks into mpg123 as a group, and filter it through sox as
   a single stream.

To this end, selector.py now detects when a track is played that has
following tracks, and returns them to nextrak as a group. This also
entails a change in nextrak itself: postSelect coming out of the
selector can now return a list of tracks, in which case nextrak will
output them to stdout as a list. (If postSelect returns nothing, the
originally selected track will be output as usual)

Since ezstream has no concept of the playlist script returning multiple
files to stream, this information is communicated to the recode script
in a different manner: nextrak-kludge will write the output to a file
called playlist.txt, which recode will read, and decode the tracks
given.

As of this moment, ezstream itself only receives the first track given
by nextrak (which, for reasons discussed previously, is always a symlink
to current.mp3). As such, it will show that track as the track title for
the entire group. This will be fixed in an update, with a (new) separate
metadata script