launch script made more robust
launch now individually checks whether icecast and ezstream are already
running, and starts or stops them as needed. Previously, if one died,
launch considered the server to be not-running, and would start a
duplicate instance and not stop previous ones.

Furthermore, it now redirects stdout and stderr into files in log/,
instead of just leaving them to dump to the console.