Moved a lot of default selector code into nextrak
The selector is no longer a module, but a class. The base class is
defined in nextrak itself, and most of the basic functionality has now
been moved into it.

selector.py now subclasses this base class, and should contain the
actual 'customization'. It contains a couple basic customizations which
should demonstrate the basic idea.